refactor(validation): move validation to a shared module
