+++ src/projects/domain/mod.rs
@@ -1,3 +1,2 @@
 pub mod name;
 pub mod project;
-pub mod validation;
