+++ src/shared/mod.rs
@@ -1,0 +1,2 @@
+//! The shared module is a storage for non domain-specific shared functionality.
+pub mod validation;
